+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -16,6 +16,7 @@ import com.wynntils.models.activities.quests.QuestModel;
 import com.wynntils.models.beacons.BeaconModel;
 import com.wynntils.models.character.CharacterModel;
 import com.wynntils.models.character.CharacterSelectionModel;
+import com.wynntils.models.character.SkillPointModel;
 import com.wynntils.models.characterstats.CharacterStatsModel;
 import com.wynntils.models.characterstats.CombatXpModel;
 import com.wynntils.models.characterstats.ShinyModel;
@@ -43,7 +44,6 @@ import com.wynntils.models.players.PlayerModel;
 import com.wynntils.models.profession.ProfessionModel;
 import com.wynntils.models.rewards.RewardsModel;
 import com.wynntils.models.seaskipper.SeaskipperModel;
-import com.wynntils.models.skillpoint.SkillPointModel;
 import com.wynntils.models.spells.SpellModel;
 import com.wynntils.models.stats.StatModel;
 import com.wynntils.models.statuseffects.StatusEffectModel;
