+++ common/src/main/java/com/wynntils/models/character/type/SavableSkillPointSet.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.skillpoint;
+package com.wynntils.models.character.type;
 
 import java.util.ArrayList;
 import java.util.Collections;
