+++ common/src/main/java/com/wynntils/screens/skillpointloadouts/SkillPointLoadoutsScreen.java
@@ -7,8 +7,8 @@ package com.wynntils.screens.skillpointloadouts;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
+import com.wynntils.models.character.type.SavableSkillPointSet;
 import com.wynntils.models.elements.type.Skill;
-import com.wynntils.models.skillpoint.SavableSkillPointSet;
 import com.wynntils.screens.base.WynntilsGridLayoutScreen;
 import com.wynntils.screens.base.widgets.TextInputBoxWidget;
 import com.wynntils.screens.base.widgets.WynntilsButton;
