+++ common/src/main/java/com/wynntils/screens/skillpointloadouts/widgets/LoadoutWidget.java
@@ -7,8 +7,8 @@ package com.wynntils.screens.skillpointloadouts.widgets;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
+import com.wynntils.models.character.type.SavableSkillPointSet;
 import com.wynntils.models.elements.type.Skill;
-import com.wynntils.models.skillpoint.SavableSkillPointSet;
 import com.wynntils.screens.skillpointloadouts.SkillPointLoadoutsScreen;
 import com.wynntils.utils.colors.CommonColors;
 import com.wynntils.utils.render.FontRenderer;
